third_party/blink/web_tests/fast/spatial-navigation/snav-single-select-list-expected.txt

	2	
4	
1
2
3
4
	6
	8	
PASS successfullyParsed is true

TEST COMPLETE

PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "8"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "2"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "6"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "4"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.getElementById("start").options[0].selected is true
PASS gFocusedDocument.getElementById("start").options[1].selected is false
PASS gFocusedDocument.getElementById("start").options[2].selected is false
PASS gFocusedDocument.getElementById("start").options[3].selected is false
PASS gFocusedDocument.getElementById("start").options[0].selected is false
PASS gFocusedDocument.getElementById("start").options[1].selected is true
PASS gFocusedDocument.getElementById("start").options[2].selected is false
PASS gFocusedDocument.getElementById("start").options[3].selected is false
PASS gFocusedDocument.getElementById("start").options[0].selected is false
PASS gFocusedDocument.getElementById("start").options[1].selected is true
PASS gFocusedDocument.getElementById("start").options[2].selected is false
PASS gFocusedDocument.getElementById("start").options[3].selected is false
PASS gFocusedDocument.getElementById("start").options[0].selected is false
PASS gFocusedDocument.getElementById("start").options[1].selected is false
PASS gFocusedDocument.getElementById("start").options[2].selected is false
PASS gFocusedDocument.getElementById("start").options[3].selected is true
PASS gFocusedDocument.getElementById("start").options[0].selected is false
PASS gFocusedDocument.getElementById("start").options[1].selected is false
PASS gFocusedDocument.getElementById("start").options[2].selected is false
PASS gFocusedDocument.getElementById("start").options[3].selected is true
PASS gFocusedDocument.getElementById("start").options[0].selected is false
PASS gFocusedDocument.getElementById("start").options[1].selected is true
PASS gFocusedDocument.getElementById("start").options[2].selected is false
PASS gFocusedDocument.getElementById("start").options[3].selected is false
PASS gFocusedDocument.getElementById("start").options[0].selected is false
PASS gFocusedDocument.getElementById("start").options[1].selected is true
PASS gFocusedDocument.getElementById("start").options[2].selected is false
PASS gFocusedDocument.getElementById("start").options[3].selected is false
PASS gFocusedDocument.getElementById("start").options[0].selected is true
PASS gFocusedDocument.getElementById("start").options[1].selected is false
PASS gFocusedDocument.getElementById("start").options[2].selected is false
PASS gFocusedDocument.getElementById("start").options[3].selected is false
PASS gFocusedDocument.getElementById("start").options[0].selected is false
PASS gFocusedDocument.getElementById("start").options[1].selected is true
PASS gFocusedDocument.getElementById("start").options[2].selected is false
PASS gFocusedDocument.getElementById("start").options[3].selected is false
